# feat(unmarshaler): add support for fields to unmarshal themselves