--- a/unmarshaler.go
+++ b/unmarshaler.go
@@ -17,6 +17,12 @@ import (
 	"github.com/pelletier/go-toml/v2/unstable"
 )
 
+// The Unmarshaler interface may be implemented by types to customize their
+// behavior when being unmarshaled from a TOML document.
+type Unmarshaler interface {
+	UnmarshalTOML(value *unstable.Node) error
+}
+
 // Unmarshal deserializes a TOML document into a Go value.
 //
 // It is a shortcut for Decoder.Decode() with the default options.
@@ -632,6 +638,12 @@ func (d *decoder) tryTextUnmarshaler(node *unstable.Node, v reflect.Value) (bool
 func (d *decoder) handleValue(value *unstable.Node, v reflect.Value) error {
 	for v.Kind() == reflect.Ptr {
 		v = initAndDereferencePointer(v)
+	}
+
+	if v.CanAddr() && v.Addr().CanInterface() {
+		if outi, ok := v.Addr().Interface().(Unmarshaler); ok {
+			return outi.UnmarshalTOML(value)
+		}
 	}
 
 	ok, err := d.tryTextUnmarshaler(value, v)