@pkarakal 2024-01-20
> Thank you for the pull request!
> 
> At the moment I think this interface should live in `unstable`. The argument to `UnmarshalTOML` is a `*unstable.Node`, which would bring `unstable` into the main (stable) toml package. This is also a partial implementation: it does not work compound values like tables and array tables. To support it, it would need to coalesce multiple nodes.
> 
> I'm not comfortable bringing in a partial feature with a ~4-5% impact on speed. I'd rather see this behavior as opt-in until it is fully implemented and the overall performance impact is close to <1% when not using the interface. Maybe an `unstable.Decoder`, which has the same API as `toml.Decoder` but with an extra `EnableUnmarshaler()` function or something like that would allow people to opt-in?

Thanks for the feedback. I will try to scratch something up, along these lines. In the meantime I will mark this as draft