@pelletier 2024-01-31
Re: making the behavior opt-in:

> Maybe an unstable.Decoder, which has the same API as toml.Decoder but with an extra EnableUnmarshaler() function or something like that would allow people to opt-in?

As of #923 I've added a precedent to put knobs that should not be considered stable as part of the public `v2` API: https://github.com/pelletier/go-toml/blob/2e087bdf5f676bd8c3cf0ea9ed3e8bf1fdc6a72c/marshaler.go#L95-L97

I suspect you would be facing similar issues I did trying to create a `unmarshal.Decoder` (it would involve moving half the lib to `internal/`). Don't worry about this, feel free to make it opt-in with a new method on the `toml.Decoder` to turn the behavior on/off. As long as the method has a disclaimer like `SetMarshalJsonNumbers` does, I'm happy with it.