@rszyma 2024-03-14
I made a PR continues work on this. Addressed the concerns about using `unstable.Node`